requirements.txt
requests==2.22.0
Markdown==3.2.2
python-dateutil==2.8.1